requirements.txt
asgiref==3.4.1
autopep8==1.6.0
chardet==4.0.0
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.8
django-excel==0.0.10
django-import-export==2.6.1
django-phonenumber-field==5.2.0
django-widget-tweaks==1.4.8
et-xmlfile==1.1.0
lml==0.1.0
MarkupPy==1.14
mysqlclient==2.0.3
odfpy==1.4.1
openpyxl==3.0.9
phonenumbers==8.12.35
Pillow==8.4.0
pycodestyle==2.8.0
pyexcel==0.6.7
pyexcel-io==0.6.5
pyexcel-webio==0.1.4
pytz==2021.3
PyYAML==6.0
sqlparse==0.4.2
tablib==3.1.0
texttable==1.6.4
toml==0.10.2
typing-extensions==3.10.0.2
xlwt==1.3.0
